feat: store out paths in decoded images
This is a bit of an incomplete implementation but the idea is to store
the decoded result of an image in it's respective task element in the
queue.
Once it has been stored we can check if it has failed or not and check
later on.
Then we obtain the paths of the decoded images