refactor: prefer parallel iterators over a thread pool
From my testing it seems that using par_iter() is a much better
implementation of multi-threading than using a threadpool.

Processing large batches of images using the thread pool causes the
program to get killed by, on my system, earlyoom, which does not happen
when running with parallel iterators.

However, threadpools can be a bit faster, since we can specify the
number of threads, we do not need them though